% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSA2D_wrapper.R
\name{ks_map}
\alias{ks_map}
\title{kinase_substrate mapping function}
\usage{
ks_map(
  protData_filename,
  psiteData_filename,
  ksNetwork_filename,
  ks_outputName
)
}
\arguments{
\item{protData_filename}{absolute directory to the file of the proteome data, in the format of the tidied up FragPipe TMT-I output}

\item{psiteData_filename}{absolute directory to the file of the single site level phospho data, in the format of the tidied up FragPipe TMT-I output, normalized or not normalized}

\item{ksNetwork_filename}{absolute directory to the kinase substrate network file, as an example, merge_ks_omni.tsv.}

\item{ks_outputName}{the file name of the output file of this function}
}
\description{
map the knase and substrate information
}
