% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSA2D_wrapper.R
\name{ksa2d_result_cytoscape}
\alias{ksa2d_result_cytoscape}
\title{generate node and edge file for cytoscape visualization}
\usage{
ksa2d_result_cytoscape(
  working_dir,
  output_dir,
  freq_cutoff = NULL,
  fdr_cutoff = 0.05,
  kinase_abs_fc_cutoff,
  subsite_abs_fc_cutoff,
  ksNetwork_filename
)
}
\arguments{
\item{working_dir}{the directory of the files that KSA2D main results are in}

\item{output_dir}{the directory for the output files of this function}

\item{freq_cutoff}{an integer indicating the number of times a kinase to substrate relationship is evaluated across all slices, default to half of the number of total slices}

\item{fdr_cutoff}{the cutoff below which kinase substrate relationships are seen as significant, default to 0.05}

\item{kinase_abs_fc_cutoff}{a positive numerical value indicating the absolute fold change cutoff of kinase}

\item{subsite_abs_fc_cutoff}{a positive numerical value indicating the absolute fold change cutoff of substrate site}

\item{ksNetwork_filename}{absolute directory to the kinase substrate network file, as an example, merge_ks_omni.tsv.}
}
\description{
visualization preparation
}
